Refactor test file locations and clean up code comments
This commit includes a restructuring of the test files whereby they have been renamed and moved to a 'http-client' directory for better organization. Additionally, verbose comments in the 'BlogHealthIndicator', 'HealthFunction', and 'StatusCodeSupplier' classes have been tidied up. A minor change in 'HealthFunctionSpec.groovy' now ensures that the health details 'HTTP Status Code' is returned as an int.

+++ src/test/groovy/lv/id/jc/piglatin/actuator/HealthFunctionSpec.groovy
@@ -29,6 +29,6 @@ class HealthFunctionSpec extends Specification {
 
         then:
         health.status == Health.down().build().status
-        health.details.get("HTTP Status Code") == httpStatusCode
+        health.details.get("HTTP Status Code") as int == httpStatusCode
     }
 }
